Edit the instruction for authentication with Matrix OpenID
Partially based on https://github.com/mother-of-all-self-hosting/mash-playbook/commit/699b458733b0d890588286a0345b3636c79f7b58

It is sensible to make it clear that Matrix User Verification Service will not be installed by this role and it is necessary to configure it outside of it. Strictly the variable used on MDAD project (matrix_user_verification_service_enabled) is not specific to this role, and it is not the task of this document to explain how UVS should be configured.

Signed-off-by: Suguru Hirahara <acioustick@noreply.codeberg.org>