Avoid specifying some jitsi-web environment variables if we have a `null` value for them
It seems like this is critical for the `RESOLUTION` environment
variable. When set to `null`, Jitsi prevents people from starting their
camera because "the required resolution is not supported by the camera"
(or something like that).

Letting the environment variable (e.g. `RESOLUTION`) be undefined
allows Jitsi to properly use its own internally-hardcoded default value
(e.g. `720`), which is better.

Ref: https://github.com/jitsi/docker-jitsi-meet/blob/31a14cc7bfdfc687887168a034584154e9fba35e/web/rootfs/defaults/settings-config.js#L36